backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.0
python-multipart==0.0.9
openpyxl==3.1.2
pandas==2.2.2
pdfplumber==0.11.0
pytesseract==0.3.10
xlsxwriter==3.2.0
weasyprint==62.1
# AI Providers
anthropic>=0.49.0
openai>=1.30.0
google-generativeai>=0.7.0
# Encryption
cryptography>=42.0.0
# Auth & Security
supabase==2.4.6
python-jose==3.3.0
slowapi==0.1.9
python-magic==0.4.27
# Background jobs
celery==5.4.0
redis==5.0.4
# Testing
pytest==8.2.0
httpx==0.27.0
